pyproject.toml
[tool.poetry]
name = "superlead-scanner"
version = "0.0.2"
description = "scan barcodes with superlead barcode reader"
authors = ["Adam Guillot <adam.guillot@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
black = {version = "^22.1.0", allow-prereleases = true}


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
